Migrations executed

+++ db/migrate/20221208010005_create_trips.rb
@@ -2,7 +2,7 @@ class CreateTrips < ActiveRecord::Migration[6.1]
   def change
     create_table :trips do |t|
       t.string :destination
-      t.integer :cost
+      t.float :cost
       t.date :check_in
       t.date :check_out
       t.integer :user_id
